+++ x/marker/keeper/genesis.go
@@ -28,6 +28,17 @@ func (k Keeper) InitGenesis(ctx sdk.Context, data *types.GenesisState) {
 	// if any markers were included directly, add these as well.
 	if data.Markers != nil {
 		for i := range data.Markers {
+			// marker base account may already exist and have an account number assigned
+			if exists := k.authKeeper.GetAccount(ctx, data.Markers[i].GetAddress()); exists != nil {
+				if err := data.Markers[i].SetAccountNumber(exists.GetAccountNumber()); err != nil {
+					panic(err)
+				}
+			} else {
+				// no existing account reference so take next account number.
+				if err := data.Markers[i].SetAccountNumber(k.authKeeper.GetNextAccountNumber(ctx)); err != nil {
+					panic(err)
+				}
+			}
 			k.SetMarker(ctx, &data.Markers[i])
 		}
 	}
@@ -40,18 +51,21 @@ func (k Keeper) ExportGenesis(ctx sdk.Context) (data *types.GenesisState) {
 	markers := make([]types.MarkerAccount, 0)
 
 	appendToMarkers := func(marker types.MarkerAccountI) bool {
-		markers = append(markers, *types.NewMarkerAccount(
-			&authtypes.BaseAccount{
+		markers = append(markers, types.MarkerAccount{
+			BaseAccount: &authtypes.BaseAccount{
 				Address:       marker.GetAddress().String(),
 				AccountNumber: marker.GetAccountNumber(),
 				Sequence:      0,
 			},
-			marker.GetSupply(),
-			marker.GetManager(),
-			marker.GetAccessList(),
-			marker.GetStatus(),
-			marker.GetMarkerType(),
-		))
+			Manager:                marker.GetManager().String(),
+			AccessControl:          marker.GetAccessList(),
+			Status:                 marker.GetStatus(),
+			Denom:                  marker.GetDenom(),
+			Supply:                 marker.GetSupply().Amount,
+			MarkerType:             marker.GetMarkerType(),
+			SupplyFixed:            marker.HasFixedSupply(),
+			AllowGovernanceControl: marker.HasGovernanceEnabled(),
+		})
 		return false
 	}