+++ x/marker/keeper/keeper.go
@@ -58,7 +58,7 @@ type Keeper struct {
 	cdc codec.BinaryCodec
 }
 
-// NewKeeper returns a metadata keeper. It handles:
+// NewKeeper returns a marker keeper. It handles:
 // - managing MarkerAccounts
 // - enforcing permissions for marker creation/deletion/management
 //